refactor: move source files to a src/ directory
Separates the configuration/setup files and
middleware code from the source files.
