+++ portal/server/app/route.ts
@@ -5,11 +5,11 @@ import { getDomain, getSubdomainAndPath } from "@lib/domain_parsing";
 import { redirectToAggregatorUrlResponse, redirectToPortalURLResponse } from "@lib/redirects";
 import { getBlobIdLink, getObjectIdLink } from "@lib/links";
 
-import { isAllowed } from "allowlist_checker";
+import { isAllowed } from "src/allowlist_checker";
 import { siteNotFound } from "@lib/http/http_error_responses";
-import integrateLoggerWithSentry from "sentry_logger";
-import blocklistChecker from "blocklist_checker";
-import { config } from "configuration_loader";
+import integrateLoggerWithSentry from "src/sentry_logger";
+import blocklistChecker from "src/blocklist_checker";
+import { config } from "src/configuration_loader";
 import { standardUrlFetcher, premiumUrlFetcher } from "url_fetcher_factory";
 import { NextRequest } from "next/server";
 import { send_to_web_analytics } from "web_analytics";
