+++ portal/server/sentry.edge.config.ts
@@ -7,7 +7,7 @@
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
 import * as Sentry from "@sentry/nextjs";
-import { config } from "configuration_loader";
+import { config } from "src/configuration_loader";
 
 if (config.enableSentry) {
     Sentry.init({
